chore: hoist state manager module
