+++ packages/@lwc/integration-karma/scripts/karma-configs/test/base.js
@@ -31,6 +31,7 @@ const ARIA_REFLECTION = require.resolve('@lwc/aria-reflection/dist/index.js');
 
 const TEST_UTILS = require.resolve('../../../helpers/test-utils');
 const TEST_SETUP = require.resolve('../../../helpers/test-setup');
+const TEST_STATE = require.resolve('../../../helpers/test-state');
 
 const ALL_FRAMEWORK_FILES = [SYNTHETIC_SHADOW, LWC_ENGINE, WIRE_SERVICE, ARIA_REFLECTION];
 
@@ -56,6 +57,7 @@ function getFiles() {
     return [
         ...frameworkFiles,
         createPattern(TEST_UTILS),
+        createPattern(TEST_STATE),
         createPattern('**/*.spec.js', { watched: false }),
     ];
 }
