+++ packages/@lwc/integration-karma/test/component/LightningElement.connectedCallback/x/contextParent/contextParent.js
@@ -1,5 +1,5 @@
 import { LightningElement } from 'lwc';
-import { nameStateFactory } from 'x/state';
+import { nameStateFactory } from 'test-state';
 
 export default class ContextParent extends LightningElement {
     random = nameStateFactory();
